ai-models-discoverer_v3/openrouter_pipeline/requirements.txt
# OpenRouter Pipeline Dependencies

# Core HTTP and web scraping
requests>=2.32.0
httpx>=0.28.0
beautifulsoup4>=4.13.0
lxml>=6.0.0

# Data processing
pandas>=2.3.0
numpy>=2.3.0

# Database and API
supabase>=2.18.0
psycopg2-binary>=2.9.9
huggingface-hub>=0.20.0

# Configuration and environment
python-dotenv>=1.1.0

# Validation and typing
pydantic>=2.11.0
typing-extensions>=4.15.0

# Date and time handling
python-dateutil>=2.9.0
pytz>=2025.2

# Security and encoding
PyJWT>=2.10.0
certifi>=2025.8.0
charset-normalizer>=3.4.0

# HTTP core dependencies
httpcore>=1.0.0
h11>=0.16.0
anyio>=4.10.0
idna>=3.10
urllib3>=2.5.0

# Utilities
six>=1.17.0
packaging>=25.0
soupsieve>=2.8
websockets>=15.0.0